feat(ring buffer): adding functionality to ring buffer. - adding is empty funtion. - adding is full function. - adding reset function. - checking emptiness on get function. - adding error return in put function.

+++ Drivers II/Core/Inc/ring_buffer.h
@@ -26,7 +26,7 @@ uint8_t ring_buffer_put(ring_buffer_t *ring_buffer,uint8_t data);
 
 uint8_t ring_buffer_get(ring_buffer_t *ring_buffer,uint8_t *data);
 
-uint8_t ring_buffer_size(ring_buffer_t *ring_buffer);
+uint16_t ring_buffer_size(ring_buffer_t *ring_buffer);
 
 uint8_t ring_buffer_is_empty(ring_buffer_t *ring_buffer);
